refactor: 도메인 계층 구조 변경
