--- android/app/src/main/java/com/on/turip/domain/videoinfo/TravelCourseRepository.kt
@@ -1,5 +1,0 @@
-package com.on.turip.domain.videoinfo
-
-interface TravelCourseRepository {
-    fun loadTripInfo(contentId: Long): Result<Trip>
-}
