--- android/app/src/main/java/com/on/turip/domain/videoinfo/contents/VideoInformation.kt
@@ -1,8 +1,0 @@
-package com.on.turip.domain.videoinfo.contents
-
-import com.on.turip.domain.videoinfo.Trip
-
-data class VideoInformation(
-    val content: Content,
-    val trip: Trip,
-)
