+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/video/trip/Place.kt
@@ -1,4 +1,4 @@
-package com.on.turip.domain.videoinfo
+package com.on.turip.domain.videoinfo.contents.video.trip
 
 data class Place(
     val placeId: Long,
