+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/video/trip/TripDuration.kt
@@ -1,4 +1,4 @@
-package com.on.turip.domain.videoinfo.contents
+package com.on.turip.domain.videoinfo.contents.video.trip
 
 data class TripDuration(
     val nights: Int,
