+++ android/app/src/main/java/com/on/turip/domain/videoinfo/contents/video/trip/repository/TravelCourseRepository.kt
@@ -1,0 +1,7 @@
+package com.on.turip.domain.videoinfo.contents.video.trip.repository
+
+import com.on.turip.domain.videoinfo.contents.video.trip.Trip
+
+interface TravelCourseRepository {
+    fun loadTripInfo(contentId: Long): Result<Trip>
+}
